Quick Sort Algorithm in Java
Quick Sort is remarkable primarily as it is the fastest known sorting algorithm. For
optimum efficiency the choice of pivot point is quite important. It is interesting to note
that as the array size increases so too does the probability of choosing an optimum pivot
point. This demonstrates that Quick Sort has the wonderful characteristic of increasing
in efficiency as it scales up to bigger data sets.